Change cache prefix only with server upgrades but not installing apps
Upgrading the OC server will clear all the caches and change the cache
prefix. This commit will add the "clear" step after a successful upgrade
to avoid leaving unused data. Changing the cache prefix was already
being done in this case.

Installing a new app won't make the cache prefix to change. This fixes a
problem installing an app using openID auth. The openID app uses a
distributed cache whose prefix was changing after installing a new app
through the market app, causing the user to logout due to wrong or
missing info (coming from the cache)

Apps are now responsible to clear their caches after the app is
upgraded, if applicable. The apps can also decide not to clean the cache